+++ process_data.go
@@ -28,7 +28,7 @@ func processLDAPUsers() {
 		}
 	}
 
-	logger(1, "LDAP User Data Processed \n", true)
+	logger(1, "LDAP Users Processed: "+fmt.Sprintf("%d", len(ldapUsers))+"\n", true)
 }
 func processData() {
 	logger(1, "Processing User Data", true)
@@ -41,6 +41,8 @@ func processData() {
 		hornbillUserData := HornbillCache.Users[userID]
 
 		if userID == "" {
+			CounterInc(7)
+			logger(4, "LDAP Record Has no User ID: "+fmt.Sprintf("%+v", currentUser.LDAP)+"\n", false)
 			continue
 		}
 		//-- Check Map no need to loop
@@ -74,7 +76,7 @@ func processData() {
 
 		logger(1, "User: "+fmt.Sprintf("%s", userID)+" \n\tCreate: "+fmt.Sprintf("%t", currentUser.Jobs.create)+" \n\tUpdate: "+fmt.Sprintf("%t", currentUser.Jobs.update)+" \n\tUpdate Type: "+fmt.Sprintf("%t", currentUser.Jobs.updateType)+" \n\tUpdate Profile: "+fmt.Sprintf("%t", currentUser.Jobs.updateProfile)+" \n\tUpdate Site: "+fmt.Sprintf("%t", currentUser.Jobs.updateSite)+" \n\tRoles Count: "+fmt.Sprintf("%d", len(currentUser.Roles))+" \n\tUpdate Image: "+fmt.Sprintf("%t", currentUser.Jobs.updateImage)+" \n\tGroups: "+fmt.Sprintf("%d", len(currentUser.Groups))+"\n", false)
 	}
-	logger(1, "User Data Processed", true)
+	logger(1, "User Data Processed: "+fmt.Sprintf("%d", len(HornbillCache.UsersWorking))+"", true)
 }
 func checkUserNeedsOrgRemoving(importData *userWorkingDataStruct, currentData userAccountStruct) {
 	for _, group := range importData.Groups {